Add reference to PhysFS library (#22)
This is a great little library that is used frequently in game dev projects. Allows a developer to get individual or groupings of files that are distributed across different archives without needing to know which archives they are in.  (also, this has nothing to do with physics even though it in the name)